+++ livereload/watcher.py
@@ -19,6 +19,9 @@ class Watcher(object):
         self._tasks = {}
         self._mtimes = {}
 
+        # setting changes
+        self._changes = []
+
         # filepath that is changed
         self.filepath = None
         self._start = time.time()
@@ -28,25 +31,39 @@ class Watcher(object):
         _, ext = os.path.splitext(filename)
         return ext in ['.pyc', '.pyo', '.o', '.swp']
 
-    def watch(self, path, func=None):
-        """Add a task to watcher."""
-        self._tasks[path] = func
+    def watch(self, path, func=None, delay=0):
+        """Add a task to watcher.
+
+        :param path: a filepath or directory path or glob pattern
+        :param func: the function to be executed when file changed
+        :param delay: delay the execution at a certain seconds
+        """
+        self._tasks[path] = (func, delay)
 
     def start(self, callback):
-        """Start the watcher running, calling callback when changes are observed. If this returns False,
-        regular polling will be used."""
+        """Start the watcher running, calling callback when changes are
+        observed. If this returns False, regular polling will be used."""
         return False
 
     def examine(self):
         """Check if there are changes, if true, run the given task."""
+        if self._changes:
+            return self._changes.pop()
+
         # clean filepath
         self.filepath = None
+        delays = set()
         for path in self._tasks:
             if self.is_changed(path):
-                func = self._tasks[path]
-                # run function
+                func, delay = self._tasks[path]
                 func and func()
-        return self.filepath
+                delays.add(delay)
+
+        if delays:
+            delay = max(delays)
+        else:
+            delay = None
+        return self.filepath, delay
 
     def is_changed(self, path):
         if os.path.isfile(path):
@@ -109,11 +126,11 @@ class INotifyWatcher(Watcher):
         self.notifier = None
         self.callback = None
 
-    def watch(self, path, func=None):
+    def watch(self, path, func=None, delay=None):
         import pyinotify
         flag = pyinotify.IN_CREATE | pyinotify.IN_DELETE | pyinotify.IN_MODIFY
         self.wm.add_watch(path, flag, rec=True, do_glob=True, auto_add=True)
-        Watcher.watch(self, path, func)
+        Watcher.watch(self, path, func, delay)
 
     def inotify_event(self, event):
         self.callback()